+++ frontend/src/router.ts
@@ -82,6 +82,12 @@ const routes = [
 		beforeEnter: validateVisit,
 		component: () => import("@/pages/PagePreview.vue"),
 	},
+	{
+		path: "/persona-survey",
+		name: "persona-survey",
+		beforeEnter: validateVisit,
+		component: () => import("@/pages/PagePersonaSurvey.vue"),
+	},
 ];
 
 declare global {
